Adds 4.22

+++ implementations/Gitter.py
@@ -29,7 +29,7 @@ def _three_elem_restriction(elems: list[float]):
         assert stufenindex_l >= 0
         n_l = N_l(stufenindex_l)
         n_l_minus_1 = N_l(stufenindex_l - 1)
-        x = np.zeros((n_l_minus_1, n_l))
+        x = np.zeros((n_l_minus_1, n_l), dtype=np.float64)
         for i in range(n_l_minus_1):
             for j in range(3):
                 x[i][2 * i + j] = elems[j]
@@ -45,7 +45,7 @@ def _three_elem_prolongation(elems: list[float]):
         assert stufenindex_l >= 0
         n_l = N_l(stufenindex_l)
         n_l_minus_1 = N_l(stufenindex_l - 1)
-        x = np.zeros((n_l, n_l_minus_1))
+        x = np.zeros((n_l, n_l_minus_1), dtype=np.float64)
         for i in range(n_l_minus_1):
             for j in range(3):
                 x[2 * i + j][i] = elems[j]
